update item detail and item search, fix some problems
