+++ src/main/resources/templates/itemDetail.html
@@ -1,33 +1,31 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns:th="https://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title>商品详情</title>
 </head>
 <body>
-<h2>别想了 输出都转义了 CSRF都加Token了</h2>
-<input type="hidden" value="这里真的没有sql注入">
 <h1>商品详情</h1>
 <table border="1">
     <tr>
         <td>商品名称</td>
-        <td th:text="${itemName}">row 1, cell 2</td>
+        <td th:text="${showItem.name}">row 1, cell 2</td>
     </tr>
     <tr>
         <td>商品价格</td>
-        <td th:text="${itemPrice}">row 2, cell 2</td>
+        <td th:text="${showItem.price}">row 2, cell 2</td>
     </tr>
     <tr>
         <td>商品状态</td>
-        <td th:text="${itemStatus}">row 2, cell 2</td>
+        <td th:text="${showItem.status}">row 2, cell 2</td>
     </tr>
     <tr>
         <td>商品图片</td>
-        <td><img th:src="${itemImage}"></td>
+        <td><img th:src="${showItem.image}"></td>
     </tr>
     <tr>
         <td>商品描述</td>
-        <td th:text="${itemDesc}">row 2, cell 2</td>
+        <td th:text="${showItem.descr}">row 2, cell 2</td>
     </tr>
 </table>
 </body>
