+++ src/main/resources/templates/search.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns:th="https://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <title>搜索结果</title>
@@ -8,7 +8,7 @@
 <h1>搜索</h1>
 <div id="searchForm">
     <h2>精确搜索</h2>
-    <form action="/api/search" method="post">
+    <form th:action="@{/api/search}" method="post">
         <input type="text" name="q" placeholder="Search" aria-label="search"/>
         <input type="submit" value="Go">
     </form>
@@ -25,11 +25,11 @@
         <td>加入购物车</td>
     </tr>
     <tr th:each="searchResult:${searchResult}">
-        <td th:text="${searchResult[itemName]}">商品名称</td>
-        <td th:text="${searchResult[itemPrice]}">商品价格</td>
-        <td th:text="${searchResult[itemStatus]}">商品状态</td>
-        <td><img th:src="${searchResult[itemImage]}"></td>
-        <td th:text="${searchResult[itemDesc]}">商品描述</td>
+        <td th:text="${searchResult.name}">商品名称</td>
+        <td th:text="${searchResult.price}">商品价格</td>
+        <td th:text="${searchResult.status}">商品状态</td>
+        <td><img th:src="${searchResult.image}"></td>
+        <td th:text="${searchResult.descr}">商品描述</td>
         <td><button>加入购物车</button></td>
     </tr>
 </table>
